requirements.txt
asgiref==3.9.1
boto3==1.40.0
botocore==1.40.0
Django==5.2.4
django-storages==1.14.6
gunicorn==23.0.0
jmespath==1.0.1
packaging==25.0
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
s3transfer==0.13.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.1
urllib3==2.5.0
whitenoise==6.9.0
